django-task-reminder/api/requirements.txt
amqp==2.4.2
astroid==2.2.5
autopep8==1.4.4
billiard==3.6.0.0
celery==4.3.0
Django==2.1
django-celery-results==1.0.4
django-cors-headers==2.5.2
django-filter==2.1.0
django-filters==0.2.1
django-redis-sessions==0.6.1
django-timezone-field==3.0
django-websocket-redis==0.5.2
djangorestframework==3.9.2
gevent==1.4.0
greenlet==0.4.15
isort==4.3.17
kombu==4.5.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pylint==2.3.1
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
vine==1.3.0
wrapt==1.11.1
